requirements.txt
# Web Framework - Güncel versiyonlar
fastapi==0.115.6
uvicorn[standard]==0.32.1

# AI/ML - Google Gemini (En son stable)
google-generativeai==0.8.3

# Data Processing - Rust sorunu olmayan versiyonlar
pydantic==2.10.4
pydantic-settings==2.7.1

# Utilities
python-multipart==0.0.18
python-dotenv==1.0.1

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Optional: Grammar & Text Analysis
# textstat==0.7.3
# language-tool-python==2.7.1